Refs #32. Checkpointing progress to refactor the code organization (2).
