+++ pyrs/core/calibration_file_io.py
@@ -108,77 +108,50 @@ class ResidualStressCalibrationFile(object):
 # END-CLASS
 
 
-def check_creation_date(file_name):
+class CalibrationManager(object):
     """
-    check the create date (year, month, date) for a file
-    :except RuntimeError: if the file does not exist
-    :param file_name: 
-    :return: 
+    A class to handle all the calibration files
+    calibration shall be related to date (run cycle), wave length and etc
     """
-    checkdatatypes.check_file_name(file_name, check_exist=True)
+    def __init__(self,  calib_lookup_table_file=None):
+        """
+        initialization for calibration manager
+        :param calib_lookup_table_file: calibration table file to in order not to scan the disk and save time
+        """
+        self._cal_dict = dict()  # dict[wavelength, date] = param_dict
+                                 # param_dict[motor position] = calibrated value
 
-    # get the creation date in float (epoch time)
-    if platform.system() == 'Windows':
-        # windows not tested
-        epoch_time = os.path.getctime(file_name)
-    else:
-        # mac osx/linux
-        stat = os.stat(file_name)
-        try:
-            epoch_time = stat.st_birthtime
-        except AttributeError:
-            # We're probably on Linux. No easy way to get creation dates here,
-            # so we'll settle for when its content was last modified.
-            epoch_time = stat.st_mtime
-        # END-TRY
-    # END-IF-ELSE
+        return
 
-    # convert epoch time to a string as YYYY-MM-DD
-    file_create_time = time.localtime(epoch_time)
-    file_create_time_str = time.strftime('%Y-%m-%d', file_create_time)
+    def get_calibration(self, ipts_number, run_number):
+        """ get calibration in memory
+        :param ipts_number:
+        :param run_number:
+        :return:
+        """
+        return
 
-    return file_create_time_str
+    def locate_calibration_file(self, ipts_number, run_number):
+        """
 
+        :param ipts_number:
+        :param run_number:
+        :return:
+        """
+        return
 
-def is_calibration_dir(cal_sub_dir_name):
-    """
-    check whether the directory name is an allowed calibration directory name for HB2B
-    :param cal_sub_dir_name:
-    :return:
-    """
-    checkdatatypes.check_file_name(cal_sub_dir_name, check_exist=True, check_writable=False,
-                                   is_dir=True, description='Directory for calibration files')
+    def show_calibration_table(self):
+        """
 
-    dir_base_name = os.path.basename(cal_sub_dir_name)
-    return dir_base_name in hb2b_setup
+        :return:
+        """
 
 
-def scan_calibration_in_archive():
-    """
-    search the archive (/HFIR/HB2B/shared/CALIBRATION/) to create a table,
-    which can be used to write the calibration information file
-    :return:
-    """
-    calib_info_table = dict()
+# END-DEF-CLASS (CalibrationManager)
 
-    calib_root_dir = '/HFIR/HB2B/CALIBRATION/'
 
-    wavelength_dir_names = os.listdir(calib_root_dir)
-    for wavelength_dir in wavelength_dir_names:
-        # skip non-relevant directories
-        wavelength_dir = os.path.join(calib_root_dir, wavelength_dir)
-        if not is_calibration_dir(wavelength_dir):
-            continue
 
-        calib_info_table[wavelength_dir] = dict()
-        cal_file_names = os.listdir(wavelength_dir)
-        for cal_file_name in cal_file_names:
-            calib_date = ResidualStressCalibrationFile(cal_file_name).retrieve_calibration_date()
-            calib_info_table[wavelength_dir][calib_date] = cal_file_name
-        # END-FOR
-    # END-FOR
 
-    return calib_info_table
 
 
 def update_calibration_info_file(cal_info_file, cal_info_table, append):
@@ -249,67 +222,6 @@ def import_calibration_info_file(cal_info_file):
     return cal_info_table
 
 
-def load_mantid_mask(pixel_number, mantid_mask_xml, is_mask):
-    """ Load Mantid mask file in XML format
-    Assumption: PixelID (detector ID) starts from 0 and there is NO gap
-    :param mantid_mask_xml:
-    :param pixel_number: total pixel number
-    :return: a vector
-    """
-    checkdatatypes.check_file_name(mantid_mask_xml, True, False, False, 'Mantid XML mask file')
-    checkdatatypes.check_int_variable('(Total) pixel number', pixel_number, (1024**2, 2048**2+1))
-
-    # load file to lines
-    mask_file = open(mantid_mask_xml, 'r')
-    mask_lines = mask_file.readlines()
-    mask_file.close()
-
-    # get detector ID range line
-    det_id_line = None
-    for line in mask_lines:
-        if line.count('<detid') > 0:
-            det_id_line = line.strip()
-            break
-    # END-FOR
-
-    if det_id_line is None:
-        raise RuntimeError('Mask file {} does not have masked detector IDs'.format(mantid_mask_xml))
-
-    # parse
-    masked_det_pair_list = det_id_line.split('>')[1].split('<')[0].strip().split(',')
-    # print ('[DB...BAT] Masked detectors range: {}'.format(masked_det_pair_list))
-
-    # create vector with 1 (for not masking)
-    masking_array = np.zeros((pixel_number,), 'float')
-    if is_mask:
-        # is given string are mask then default is not masked
-        masking_array += 1.
-    # is ROI default = 0
-
-    masked_specs = 0
-    for masked_det_pair in masked_det_pair_list:
-        # get range
-        terms = masked_det_pair.split('-')
-        start_detid = int(terms[0])
-        end_detid = int(terms[1])
-        # check range
-        if end_detid >= pixel_number:
-            raise RuntimeError('Detector ID {} is out of range of given detector size {}'
-                               ''.format(end_detid, pixel_number))
-        # mask or ROI
-        if is_mask:
-            masking_array[start_detid:end_detid+1] = 0.
-        else:
-            masking_array[start_detid:end_detid+1] = 1.
-        # stat
-        masked_specs += end_detid - start_detid + 1
-    # END-FOR
-
-    print ('[DB...CHECK] Masked spectra = {}, Sum of masking array = {}'
-           ''.format(masked_specs, sum(masking_array)))
-
-    return masking_array
 
 
-# testing
-print (check_creation_date('__init__.py'))
+
